requirements.txt
Django>5.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-debug-toolbar
psycopg2-binary
django-environ
dj-rest-auth
django-allauth
django-extensions
requests
cryptography
requests_oauthlib
google-api-python-client
drf-spectacular
django-countries
pillow
django-cors-headers
pandas
openpyxl
django-cities-light
django-rest-passwordreset
opencv-python
mediapipe
numpy
boto3
django-storages
stripe
daphne
firebase-admin